[spdx3.0] move core class to model level
